pythonSimplified/requirements.txt
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
Brlapi==0.8.2
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
click==8.1.2
colorama==0.4.4
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.6.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.5.0
distro-info==1.0
duplicity==0.8.20
entrypoints==0.4
executing==0.8.3
fasteners==0.14.1
fastjsonschema==2.15.3
filelock==3.6.0
findspark==2.0.1
Flask==2.1.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.18.1
idna==2.10
importlib-metadata==4.0.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.3.0
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keyring==23.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lockfile==0.12.2
louis==3.18.0
macaroonbakery==1.3.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
paramiko==2.7.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.3
pycairo==1.16.2
pycparser==2.21
pycups==2.0.1
Pygments==2.12.0
PyGObject==3.40.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyparsing==3.0.8
pyRFC3339==1.1
pyrsistent==0.18.1
pyspark==3.2.1
python-apt==2.2.1
python-dateutil==2.8.2
python-debian===0.1.39ubuntu1
pytz==2021.1
pyxdg==0.27
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
reportlab==3.5.66
requests==2.25.1
SecretStorage==3.3.1
Send2Trash==1.8.0
simplejson==3.17.2
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
stack-data==0.2.0
systemd-python==234
terminado==0.13.3
tinycss2==1.1.1
tornado==6.1
traitlets==5.1.1
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.14.1
wadllib==1.3.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.1
widgetsnbextension==3.6.0
WTForms==3.0.1
xdg==5
xkit==0.0.0
zipp==1.0.0
